+++ HISTORY.md
@@ -2,8 +2,14 @@
 Release History
 ===============
 
-0.3
+0.4
 ---
+
+- Add children user time in cpu metric
+- Add metric ``num_threads``
+
+0.3 (2019-09-09)
+----------------
 
 - Rework of the dashboard to separate value and display (new format for the description file)
 - Make the dashboard configuration file a Jinja2 file (to be able to define variables, macros, etc.)